requirement.txt
numpy
matplotlib
pdf2image
pytesseract
tesserocr
pillow